fixed imports and moved glitch files into main tod_obs directory
