+++ sotodlib/tod_ops/filter_stats_functions.py
